logging interceptor and bottom app bar scrolling behavior
