+++ app/build.gradle.kts
@@ -71,7 +71,9 @@ dependencies {
     //retrofit
     implementation("com.squareup.retrofit2:retrofit:2.9.0")
     implementation("com.squareup.retrofit2:converter-gson:2.9.0")
+    //okhttp
     implementation("com.squareup.okhttp3:okhttp:4.9.0")
+    implementation("com.squareup.okhttp3:logging-interceptor:4.9.0")
     //hilt
     implementation("com.google.dagger:hilt-android:2.48")
     implementation("androidx.compose.material3:material3-android:1.2.0")
