+++ app/src/main/java/com/ahugenb/tt/home/HomeActivity.kt
@@ -3,20 +3,12 @@ package com.ahugenb.tt.home
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.ahugenb.tt.R
-import com.ahugenb.tt.match.view.MatchesScreen
 import com.ahugenb.tt.ui.theme.TennisTimeTheme
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -34,7 +26,7 @@ class HomeActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    TennisTimeAppScreen(
+                    BottomAppBar(
                         currentScreen = mainState.currentScreen,
                         onNavigationItemSelected = viewModel::onNavigationItemSelected
                     )
@@ -44,23 +36,3 @@ class HomeActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun TennisTimeAppScreen(
-    currentScreen: Screen,
-    onNavigationItemSelected: (Screen) -> Unit
-) {
-    Scaffold(
-        bottomBar = { AppBottomNavigation(currentScreen, onNavigationItemSelected) }
-    ) { innerPadding ->
-        Box(modifier = Modifier.padding(innerPadding)) {
-            when (currentScreen) {
-                Screen.HIGHLIGHTED_PLAYER -> Text(stringResource(R.string.tab_highlighted_player))
-                Screen.TOURNAMENTS -> Text(stringResource(R.string.tab_tournaments))
-                Screen.LIVE_MATCHES -> MatchesScreen()
-                Screen.ATP_RANKINGS -> Text(stringResource(R.string.tab_atp_rankings))
-                Screen.WTA_RANKINGS -> Text(stringResource(R.string.tab_wta_rankings))
-            }
-        }
-
-    }
-}
